add benchmark for heap leaks
